code reformat

+++ DCCPanelController/View/Converters/PanelToHeightConverter.cs
@@ -5,13 +5,11 @@ namespace DCCPanelController.View.Converters;
 
 public class PanelToCardHeightConverter : IMultiValueConverter {
     public object Convert(object[] values, Type targetType, object? parameter, CultureInfo culture) {
-        
         if (values is [double cardWidth, Panel panel, ..]) {
-
             // What is the ratio of the Columns to Rows
             // --------------------------------------------------------------
             if (cardWidth > 0) {
-                var ratio = (double)panel.Rows / (double)panel.Cols;
+                var ratio = panel.Rows / (double)panel.Cols;
                 var itemHeight = cardWidth * ratio;
                 return itemHeight;
             }
@@ -22,5 +20,4 @@ public class PanelToCardHeightConverter : IMultiValueConverter {
     public object[] ConvertBack(object value, Type[] targetTypes, object parameter, CultureInfo culture) {
         throw new NotImplementedException();
     }
-
 }
